requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# MCP SDK (if available) or HTTP server
websockets==12.0
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# HTTP client for API communication
httpx==0.25.2

# Web framework and templates
jinja2==3.1.2
aiofiles==23.2.1

# Google Gemini AI
google-generativeai==0.8.3

# WebSocket for real-time chat
websockets==12.0